refactor: 필터 섹션 리팩토링
